Lab 2
Within this lab, I was tasked to create a program that reads two user-inputted values. Using these values, the program will store them in two variables named X and Y, print the values of X and Y, print the value of XY, print X/Y, print the value of (X+1) * 6, and prints the value of X and Y a secondary time to show the values never changed.