Lab 3
Within this lab, I was tasked to create a program that goes through all even integers from 1 through 100 in order from the greatest integer down to the least integer. While moving through the integers, if the current integer is divisible by 3, the program will print “T". If the current integer is divisible by 7, the program will print “S”.  If the current integer is divisible by both 7 and 3, the program will print “TS”. Following the letters, the integer itself will be printed.