Lab 4
Within this lab, I was tasked to create a program that spells the word of the integer from all integers 1-99 using a switch. For example, if the integer is equal to 2, the program will print "Two". After completing this lab, I noticed that a much more efficient approach to this task would be through integer division. For example, by dividing the integer by 10, you could determine what value (twenty, thirty, forty, etc.) would be in the tens place and using the remainder (one, two, three) within the ones place.