Pre-Lab 3
Within this pre-lab practice assignment, I was tasked to create a program that goes through all integers 1-50 beginning at 1. While moving through the integers, if the current integer is divisible by 3, the program will print “Fizz”. If the current integer is divisible by 5, the program will print “Buzz”. If the current integer is divisible by both 5 and 3, the program will print “Fizz Buzz”. If the current integer is divisible by neither 5 nor 3, the program will print the integer.