Pre-Lab 5
Within this pre-lab practice assignment, I was tasked to write a function that returns the mathematical number e as a double and prints it from main(). Additionally, I was tasked to have the program raised e to all integer powers 1 through 5. When the program is executed, it will print out the values of e to an integer power.