Pre-Lab 6
Within this pre-lab assignment, I was tasked to declare an array of integers in the main() function and initialize the elements of the array to integer values 1 through 20. Additionally, I was tasked to find the sum and average of the array. When executed, the program will print out each element of the array, the sum of all elements in the array, and the average of the elements in the array.