Pre-Lab 7
Within this pre-lab assignment, we were given a table of data regarding student and their corresponding grades for different classes. I was initially tasked to find the averages of each row (student's average grade) and column (class' average grade), however, I wanted to challenge myself and learn more about multi-dimensional arrays. Therefore, I made additional functions including adding 100 to each element in the array, finding the sums of the columns and the rows, finding the average of all elements in the array, along with reassigning every element in the array to 0.