Pre-Lab 8
Within this pre-lab practice assignment, I was tasked to implement a program that reads a one-dimensional array from a user via a function. When the program is executed, it will prompt the user asking for an integer number representing the size of the array.  It will then ask for each element in the array and once all values are inputted, it will print the entire array.