Lab 5
Within this lab, I was tasked to write a function that returns the summation of the square of numbers from 1 through the limit of n. For this specific assignment, all integer values 1 through 10 were to be put through the function and then return the result. When executed, the program will calculate the square of the integer and add it to the sum of the past squared integers. For each integer, it will print the summation.