Lab 6
Within this lab, I was tasked to create and print an array of integers that is initialized to the odd numbers from 1 through 99. Additionally, the program was to print the sum and averages of the even and odd elements. When executed, the program will print the array, calculated the sums and averages, and then return the values to main() to be printed.