Lab 7
Within this lab, we were assigned to declare two multi-dimensional arrays. One of these arrays represents students grades in multiple classes and the other was to represent the grade cutoffs for each class. After printing both arrays, the program will find and print the grade for each of the student's classes.